requirements.txt
altgraph==0.17.2
Pillow==9.2.0
pyinstaller==5.3
pyinstaller-hooks-contrib==2022.10
